add js route
