+++ assets/pages/digital-subscription-checkout/components/checkoutStage.jsx
@@ -5,10 +5,6 @@
 import React from 'react';
 import { connect } from 'react-redux';
 
-import ProductHero, {
-  type GridImages,
-  type ImagesByCountry,
-} from 'components/productHero/productHero';
 import HeadingBlock from 'components/headingBlock/headingBlock';
 import { HeroWrapper } from 'components/productPage/productPageHero/productPageHero';
 import { type CountryGroupId } from 'helpers/internationalisation/countryGroup';
@@ -21,9 +17,9 @@ import { type Stage, type State } from '../digitalSubscriptionCheckoutReducer';
 
 import ThankYouContent from './thankYouContent';
 import ThankYouPendingContent from './thankYouPendingContent';
-import ThankYouExistingContent from './thankYouExistingContent';
 import CheckoutForm from './checkoutForm';
 import ReturnSection from './returnSection';
+import ThankYouHero from './thankYou/hero';
 
 
 // ----- Types ----- //
@@ -33,68 +29,6 @@ type PropTypes = {|
   formSubmitted: boolean,
   countryGroupId: CountryGroupId,
 |};
-
-
-// ----- Setup ----- //
-
-const defaultHeroes: GridImages = {
-  breakpoints: {
-    mobile: {
-      gridId: 'digitalSubscriptionHeaderMobile',
-      srcSizes: [342, 684, 1200],
-      imgType: 'png',
-    },
-    tablet: {
-      gridId: 'digitalSubscriptionHeaderTablet',
-      srcSizes: [500, 1000, 2000],
-      imgType: 'png',
-    },
-    desktop: {
-      gridId: 'digitalSubscriptionHeaderDesktop',
-      srcSizes: [500, 1000, 2000, 4045],
-      imgType: 'png',
-    },
-  },
-  fallback: 'digitalSubscriptionHeaderDesktop',
-};
-
-const heroesByCountry: ImagesByCountry = {
-  GBPCountries: defaultHeroes,
-  UnitedStates: defaultHeroes,
-  International: defaultHeroes,
-  EURCountries: defaultHeroes,
-  NZDCountries: defaultHeroes,
-  Canada: defaultHeroes,
-  AUDCountries: {
-    breakpoints: {
-      mobile: {
-        gridId: 'digitalSubscriptionHeaderMobileAU',
-        srcSizes: [310, 620, 1088],
-        imgType: 'png',
-      },
-      tablet: {
-        gridId: 'digitalSubscriptionHeaderTabletAU',
-        srcSizes: [500, 1000, 2000],
-        imgType: 'png',
-      },
-      desktop: {
-        gridId: 'digitalSubscriptionHeaderDesktopAU',
-        srcSizes: [500, 1000, 2000, 4045],
-        imgType: 'png',
-      },
-    },
-    fallback: 'digitalSubscriptionHeaderDesktopAU',
-  },
-};
-
-const ThankYouHero = ({ countryGroupId }: {countryGroupId: CountryGroupId}) => (
-  <ProductHero
-    countryGroupId={countryGroupId}
-    imagesByCountry={heroesByCountry}
-    altText="digital subscription"
-    fallbackImgType="png"
-  />
-);
 
 
 // ----- State/Props Maps ----- //
@@ -144,23 +78,6 @@ function CheckoutStage(props: PropTypes) {
             </HeadingBlock>
           </HeroWrapper>
           <ThankYouPendingContent />
-          <ReturnSection />
-        </div>
-      );
-
-
-    case 'thankyou-existing':
-      return (
-        <div className="thank-you-stage">
-          <ThankYouHero
-            countryGroupId={props.countryGroupId}
-          />
-          <HeroWrapper appearance="custom">
-            <HeadingBlock>
-              Your Digital Pack subscription is already live
-            </HeadingBlock>
-          </HeroWrapper>
-          <ThankYouExistingContent countryGroupId={props.countryGroupId} />
           <ReturnSection />
         </div>
       );
