+++ assets/pages/digital-subscription-checkout/digitalSubscriptionCheckoutReducer.js
@@ -7,7 +7,7 @@ import { combineReducers, type Dispatch } from 'redux';
 import { type ReduxState } from 'helpers/page/page';
 import { type Option } from 'helpers/types/option';
 import { type DigitalBillingPeriod, Monthly } from 'helpers/billingPeriods';
-import { getQueryParameter, getPathAfterRoute } from 'helpers/url';
+import { getQueryParameter } from 'helpers/url';
 import csrf, { type Csrf as CsrfState } from 'helpers/csrf/csrfReducer';
 import {
   fromString,
@@ -35,7 +35,7 @@ import { showPaymentMethod, onPaymentAuthorised, countrySupportsDirectDebit } fr
 
 // ----- Types ----- //
 
-export type Stage = 'checkout' | 'thankyou' | 'thankyou-pending' | 'thankyou-existing';
+export type Stage = 'checkout' | 'thankyou' | 'thankyou-pending';
 type PaymentMethod = 'Stripe' | 'DirectDebit';
 
 export type FormFieldsInState = {|
@@ -231,10 +231,8 @@ function initReducer(initialCountry: IsoCountry) {
     : Monthly;
   const { productPrices } = window.guardian;
 
-  const checkoutPathName: string = getPathAfterRoute('checkout').pop();
-
   const initialState = {
-    stage: checkoutPathName === 'existing' ? 'thankyou-existing' : 'checkout',
+    stage: 'checkout',
     email: user.email || '',
     firstName: user.firstName || '',
     lastName: user.lastName || '',
