+++ conf/routes
@@ -93,7 +93,7 @@ GET  /$country<(uk|us|au|eu|int)>/subscribe/digital     controllers.DigitalSubsc
 # redirect from old checkout urls
 GET  /$country<(uk|us|au|int)>/subscribe/digital/checkout  controllers.Application.permanentRedirectWithCountry(country, location="/subscribe/digital/checkout")
 GET  /subscribe/digital/checkout  controllers.DigitalSubscription.displayForm()
-GET  /subscribe/digital/checkout/existing  controllers.DigitalSubscription.displayForm()
+GET  /subscribe/digital/checkout/thankyou-existing  controllers.DigitalSubscription.displayThankYouExisting()
 POST /subscribe/digital/create  controllers.DigitalSubscription.create
 
 GET  /premium-tier                            controllers.Subscriptions.premiumTierGeoRedirect()
